refactor: split long task docs and inline unit tests
Split the 435-line docs/troubleshooting.md (pure prose, zero
specdown scripts) into six topic subpages under docs/troubleshooting/,
and update the README link. Relative links were re-anchored for the
new depth.

Move the thirteen *_test.rs sibling modules in mit-commit-message-lints
into idiomatic inline `#[cfg(test)] mod tests` blocks at the foot of
the source file they test. These exercised pub(crate)/crate-internal
items so they could not move to a top-level tests/ dir; inlining keeps
them as true unit tests next to the code under test. All 71 tests and
the crate's strict clippy::pedantic config are preserved.

Fix two pre-existing clippy::manual_is_variant_and errors in git-mit
and git-mit-relates-to (.ok().is_some_and -> .is_ok_and) uncovered by
a newer nightly clippy, so `just lint` is green again.

No tests were added or removed. cargo test --workspace: 332 passed.

# Conflicts:
#	mit-commit-message-lints/src/mit/cmd/mod.rs
#	mit-commit-message-lints/src/mit/lib/mod.rs
