+++ mit-commit-message-lints/src/mit/cmd/mod.rs
@@ -3,11 +3,7 @@ const CONFIG_KEY_EXPIRES: &str = "mit.author.expires";
 const CONFIG_KEY_ROTATION: &str = "mit.author.rotate";
 
 pub(crate) mod get_authors;
-#[cfg(test)]
-mod get_authors_test;
 pub(crate) mod get_commit_coauthor_configuration;
-#[cfg(test)]
-pub(crate) mod get_commit_coauthor_configuration_test;
 pub(crate) mod set_commit_authors;
 pub(crate) mod set_config_authors;
 
@@ -19,19 +15,9 @@ pub mod get_config_non_clean_behaviour;
 /// When rotation is enabled, the primary author (user.name/user.email)
 /// rotates among the configured authors with each commit.
 pub mod get_config_rotation;
-#[cfg(test)]
-mod get_config_rotation_test;
 /// Rotate the primary author among configured authors
 pub mod rotate_authors;
-#[cfg(test)]
-mod rotate_authors_test;
-#[cfg(test)]
-mod set_commit_authors_test;
-#[cfg(test)]
-mod set_config_authors_test;
 pub mod set_config_non_clean_behaviour;
 /// Configuration for rotating primary author across commits
 pub mod set_config_rotation;
-#[cfg(test)]
-mod set_config_rotation_test;
 mod vcs;
