+++ mit-commit-message-lints/src/mit/lib/author.rs
@@ -44,3 +44,30 @@ impl<'a> Author<'a> {
         self.signingkey.as_deref()
     }
 }
+
+#[cfg(test)]
+mod tests {
+    #![allow(clippy::wildcard_imports)]
+
+    use super::*;
+
+    #[test]
+    fn test_new_author_creation() {
+        let author = Author::new("The Name".into(), "email@example.com".into(), None);
+
+        assert_eq!(author.name(), "The Name");
+        assert_eq!(author.email(), "email@example.com");
+        assert_eq!(author.signingkey(), None);
+    }
+
+    #[test]
+    fn test_author_with_signing_key() {
+        let author = Author::new(
+            "The Name".into(),
+            "email@example.com".into(),
+            Some("0A46826A".into()),
+        );
+
+        assert_eq!(author.signingkey(), Some("0A46826A"));
+    }
+}
