+++ mit-commit-message-lints/src/mit/lib/mod.rs
@@ -2,19 +2,7 @@
 
 pub(crate) mod author;
 pub(crate) mod author_state;
-#[cfg(test)]
-mod author_state_test;
-#[cfg(test)]
-pub(crate) mod author_test;
 pub(crate) mod authors;
-#[cfg(test)]
-mod authors_test;
 mod errors;
-#[cfg(test)]
-mod errors_test;
 pub mod non_clean_behaviour;
-#[cfg(test)]
-mod non_clean_behaviour_test;
 pub mod rotation_option;
-#[cfg(test)]
-mod rotation_option_test;
