+++ mit-commit-message-lints/src/relates/cmd/mod.rs
@@ -1,7 +1,3 @@
 pub mod errors;
 pub mod get_relate_to_configuration;
-#[cfg(test)]
-mod get_relate_to_configuration_test;
 pub mod set_relates_to;
-#[cfg(test)]
-mod set_relates_to_test;
