+++ mit-commit-message-lints/src/relates/lib/mod.rs
@@ -1,3 +1,1 @@
 pub mod relates_to;
-#[cfg(test)]
-mod relates_to_test;
